[Refactor][Elementwise] Read dtype at forward across the elementwise family
Closes #1832.

24 constructors stop taking `dtype`. `AlibiFwdOp` and `SinusoidalFwdOp`
keep it: they generate a tensor from their arguments and have no input to
read one from.

The family gets one way to reach a kernel. `_PerDtypeKernels._entry(dtype)`
is the single lookup; a subclass supplies `_build_entry` and gets back a
`KernelEntry` carrying everything that specialization implies — the kernel,
the dtype it was compiled for, the resolved output dtype, and whether bool
operands are reinterpreted as uint8. An op with no such split simply has
`compute_dtype == dtype`.

That record replaces three separate mechanisms: the `_bool_storage` slot
written as a side effect of building a kernel and read by two other
methods, the `_IntIdentityUnaryOp` constructor branch that skipped kernel
construction for integer dtypes, and the `output_dtype_override` passed at
registration — the manifest already answers that, including the integer
promotion an override could not express.

`self.dtype` is assigned in one place, inside `_entry`. It is metadata for
`eval_roofline` and `total_memory`, never read on an execution path: by the
time a second dtype arrives it no longer describes the call in flight.
Before this it was assigned in 21 places, and `RoundFwdOp` read a stale one
— an integer call followed by a float call skipped rounding entirely.

Scalar parameters are baked into kernels, so they can only be checked
against an element type. Those checks move into the entry builders. The
visible consequence is timing: an out-of-range scalar now raises at first
use. Tests that asserted a rejection at construction assert it at the call,
and those that asserted "input dtype must equal the declared dtype" now
assert what survives — every dtype in the manifest union is accepted by one
instance, and a dtype outside it still raises.

Deletes `_FROZEN_UNARY_ACTIVATION_SIGNATURES`, a hand-copied snapshot of ten
constructor signatures. The validator's C3 check enforces the same contract
against the manifest, so the table only ever fired when the contract changed
on purpose.

+++ benchmarks/ops/bench_independent_elementwise.py
@@ -227,7 +227,7 @@ def _sinusoidal_reference(seq_len: int, d_model: int, dtype: torch.dtype) -> tor
 
 @AlibiBenchFixture
 def test_alibi_bench(seq_len: int, num_heads: int, dtype: torch.dtype) -> None:
-    op = AlibiFwdOp(seq_len=seq_len, num_heads=num_heads)
+    op = AlibiFwdOp(seq_len=seq_len, num_heads=num_heads, dtype=dtype)
     workload = _GenerativeWorkload((num_heads, seq_len, seq_len), dtype)
     bm = ManifestBenchmark(_ALIBI_OP, op, workload)
 
@@ -240,7 +240,7 @@ def test_alibi_bench(seq_len: int, num_heads: int, dtype: torch.dtype) -> None:
 
 @SinusoidalBenchFixture
 def test_sinusoidal_bench(seq_len: int, d_model: int, dtype: torch.dtype) -> None:
-    op = SinusoidalFwdOp(seq_len=seq_len, d_model=d_model)
+    op = SinusoidalFwdOp(seq_len=seq_len, d_model=d_model, dtype=dtype)
     workload = _GenerativeWorkload((seq_len, d_model), dtype)
     bm = ManifestBenchmark(_SINUSOIDAL_OP, op, workload)
